requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pyserial==3.5
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
websockets==12.0

# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
pytest-mock==3.12.0
